+++ R/batch-correction.R
@@ -63,7 +63,7 @@ batch_correct_counts <- function(
   label_offset_x_ = 2,
   label_offset_y_ = 2,
   label_font_size = 3,
-  point_size_for_pca = 3,
+  point_size_for_pca = 5,
   color_histogram_by_group = TRUE,
   set_min_max_for_x_axis_for_histogram = FALSE,
   minimum_for_x_axis_for_histogram = -1,
@@ -121,6 +121,7 @@ batch_correct_counts <- function(
   if (is.null(samples_to_include)) {
     samples_to_include <- sample_metadata |> dplyr::pull(sample_id_colname)
   }
+  pca_label_colname <- if (isTRUE(add_label_to_pca)) label_colname else NULL
   if (is.null(label_colname)) {
     label_colname <- sample_id_colname
   }
@@ -179,7 +180,7 @@ batch_correct_counts <- function(
       sample_id_colname = sample_id_colname,
       feature_id_colname = feature_id_colname,
       group_colname = batch_colname,
-      label_colname = label_colname,
+      label_colname = pca_label_colname,
       samples_to_rename = samples_to_rename,
       color_values = colors_for_plots,
       principal_components = c(
@@ -188,7 +189,6 @@ batch_correct_counts <- function(
       ),
       legend_position = legend_position_for_pca,
       point_size = point_size_for_pca,
-      add_label = add_label_to_pca,
       label_font_size = label_font_size,
       label_offset_y_ = label_offset_y_,
       label_offset_x_ = label_offset_x_,
@@ -211,9 +211,12 @@ batch_correct_counts <- function(
       maximum_for_x_axis = maximum_for_x_axis_for_histogram,
       legend_position = legend_position_for_histogram,
       legend_font_size = legend_font_size_for_histogram,
-      number_of_legend_columns = number_of_histogram_legend_columns
-    ) +
-      ggplot2::labs(caption = "batch-corrected counts")
+      number_of_legend_columns = number_of_histogram_legend_columns,
+      interactive_plots = interactive_plots
+    )
+    if (!isTRUE(interactive_plots)) {
+      hist_plot <- hist_plot + ggplot2::labs(caption = "batch-corrected counts")
+    }
     if (isTRUE(plot_corr_matrix_heatmap)) {
       corHM_plot <- plot_corr_heatmap(
         combat_edata,
@@ -235,7 +238,7 @@ batch_correct_counts <- function(
 
     plot_ext <- "png"
     if (isTRUE(interactive_plots)) {
-      pca_plot <- pca_plot |> plotly::ggplotly(tooltip = c("sample", "group"))
+      pca_plot <- pca_plot |> plotly::ggplotly(tooltip = "text")
       hist_plot <- (hist_plot + ggplot2::theme(legend.position = "none")) |>
         plotly::ggplotly(tooltip = c("sample"))
       plot_ext <- "html"